requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Database
sqlalchemy==2.0.32
pymysql==1.1.1
aiomysql==0.2.0
alembic==1.13.2

# Validation
pydantic==1.10.13
email-validator

# Auth & Security
python-jose[cryptography]==3.4.0
passlib[bcrypt]>=1.7.4,<1.8
bcrypt>=3.2.0,<4.0
python-multipart>=0.0.18

# Celery + Redis
celery==5.4.0
redis==5.0.8

# HTTP Requests
requests==2.32.0

# Utils / Environment
python-dotenv==1.0.1

